requirements.txt
opencv-python
torch
facenet-pytorch
scikit-learn
numpy
pyyaml
faiss-cpu
insightface
